+++ changes.md
@@ -1,4 +1,7 @@
 ## RARL
-
 - They assume transition function has errors instead of perfect environment transitions
-- 
+
+## General
+- Make continuous algs (DDPG)
+- Generalize to mujoco envs
+- Make RARL